Separated modules into directory
