+++ demeuk/modules/remove.py
@@ -7,7 +7,7 @@
 # log is a debug string which can be None. Logged when result is True (something changed)
 from re import search, sub
 
-from add import global_store_punctuation, get_punctuation
+from modules.add import global_store_punctuation, get_punctuation
 from regexes import EMAIL_REGEX
 
 global_store_delims = [':']
